Little refactor that adds plugins/xxx/index.ts files
