+++ lib/plugins/core/index.ts
@@ -1,0 +1,5 @@
+import CorePlugin from './CorePlugin';
+
+const plugin = new CorePlugin();
+
+export { plugin };
