+++ lib/plugins/images/index.ts
@@ -1,0 +1,5 @@
+import ImagesPlugin from './ImagesPlugin';
+
+const plugin = new ImagesPlugin();
+
+export { plugin };
